+++ BeansLib/src/Beans/Withdraw_SavingRemote.java
@@ -9,5 +9,5 @@ import javax.ejb.Remote;
 @Remote
 public interface Withdraw_SavingRemote {
 
-    public void takeBalance(Integer E_ID, String ACCNUM, Integer BALANCE, String desc);
+    public void takeBalance(Integer E_ID, String ACCNUM, Double BALANCE, String desc);
 }
